+++ djangobmf/models/__init__.py
@@ -6,7 +6,6 @@ from __future__ import unicode_literals
 __all__ = (
     'BMFModel',
     'BMFModelBase',
-    'BMFModelMPTT',
     'ACTION_COMMENT',
     'ACTION_CREATED',
     'ACTION_UPDATED',
@@ -39,7 +38,6 @@ import json
 
 from .base import BMFModel
 from .base import BMFModelBase
-from .base import BMFModelMPTT
 
 from .activity import ACTION_COMMENT
 from .activity import ACTION_CREATED